Unify top level objects in extension.js
